[update] add modal when delete item

+++ src/modules/garment-finance/master-pembukuan/view.js
@@ -1,12 +1,14 @@
 import { inject, Lazy } from 'aurelia-framework';
 import { Router } from 'aurelia-router';
 import { Service } from './service';
+import { Dialog } from '../../../au-components/dialog/dialog'
 
-@inject(Router, Service)
+@inject(Router, Service, Dialog)
 export class View {
-	constructor(router, service) {
+	constructor(router, service, dialog) {
 		this.router = router;
 		this.service = service;
+		this.dialog = dialog;
 	}
 
 	async activate(params) {
@@ -23,15 +25,15 @@ export class View {
 	}
 
 	deleteCallback(event) {
-		// this.dialog.prompt('Apakah yakin menghapus data?', 'Hapus Data')
-		// 	.then(response => {
-		// 		if (response.ok) {
+		this.dialog.prompt('Apakah yakin menghapus data?', 'Hapus Data')
+			.then(response => {
+				if (response.ok) {
 					this.service.delete(this.data)
 						.then(result => {
 							this.cancelCallback();
 						});
-		// 		}
-		// 	});
+				}
+			});
 
 		// this.service.delete(this.data)
     //   .then(result => {
